fix database
schedule.time -> schedule.start_time & schedule.end_time

+++ database/migrations/2022_06_21_120015_create_attendances_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
         Schema::create('attendances', function (Blueprint $table) {
             $table->id();
             $table->foreignId('class_id')->constrained();
-            $table->bigInteger('period');
+            $table->integer('period');
             $table->date('date')->default(DB::raw('(CURRENT_DATE)'));
             $table->unique(['class_id','period']);
         });
